5/3/2022

+++ base/views.py
@@ -1,10 +1,58 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
-from .models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Message, Room, Topic
 from .forms import RoomForm
 
 
 # Create your views here.
+
+def loginPage(request):
+    page = 'login'
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist!!')
+
+        user = authenticate(request,username=username, password=password)
+
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request, 'username or password does ')
+        
+    context = {'page' : page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form  = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration')
+    return render(request, 'base/login_register.html',{'form' : form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -15,16 +63,32 @@ def home(request):
         )
 
     topics = Topic.objects.all()
-   
-    context = {'rooms' : rooms, 'topics' : topics}
+    room_count = rooms.count()
+    messages = Message.objects.all()
+    
+    context = {'rooms' : rooms, 'topics' : topics, 'room_count': room_count, 'messages' : messages}
     
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
     rooms = Room.objects.get(id=pk)
-    context = {'rooms' : rooms}
-    return render(request,'base/room.html', context)
+    messages = rooms.message_set.all().order_by('-created')
+    participants = rooms.participants.all()
 
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = rooms,
+            body = request.POST.get('body')
+        )
+        rooms.participants.add(request.user)
+        return redirect('room', pk=rooms.id)
+    
+    context = {'rooms' : rooms, 'messages' : messages, 'participants' : participants}
+    return render(request,'base/room.html', context)
+ 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -36,10 +100,14 @@ def createRoom(request):
     context = {'form' : form}  
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -49,9 +117,26 @@ def updateRoom(request, pk):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obj' : room})
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request, 'base/delete.html', {'obj' : message})
